requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
python-dotenv>=1.0
dj-database-url>=2.1